Make viewer work better with HighDPI screens (pixelDensity)
The pixelDensity(1) call was removed, so it is automatically computed by
P5.js as described in https://p5js.org/reference/#/p5/pixelDensity

Now, text is as sharp on HighDPI screens as on regular ones. Otherwise,
the text is (was) blurry. I looked a bit through the git history of your
project but could not find a reason for the pixelDensity(1) call, as it
works better without it. Maybe there was an issue with old P5.js
versions? Works perfectly on my end now though.

I hope including a new screenshot is fine for this change, it also
demonstrates how it looks without blurry text. My concern is that this
may not pass some testing pipeline now, but that has to be seen on other
setups... Interestingly, colors also look a bit more lively with the
automatically computed pixel density.